1.add lcd display mode: H/V 2.add encryption protocol

+++ bos/drivers/inc/b_ssd1289.h
@@ -55,8 +55,8 @@
  
 typedef struct
 {
-    void (*pWriteCmd)(uint8_t);
-    void (*pWriteDat)(uint8_t);
+    void (*pWriteCmd)(uint16_t);
+    void (*pWriteDat)(uint16_t);
 }bSSD1289Private_t;
